Improves GeoViewSync
This change disables the 'flick' gesture for both map and sceneview.
This gesture was problematic, because it was possible for the user to
start navigating on the scene, only to be fought when attempting to
navigate the map - the sceneview was still actively navigating.

With this change, the sample is now demonstrating similar behavior to
what is seen in ArcGIS pro when a map and a scene are linked.