Enable use of a floating form control on Forms
UWP and WPF samples follow the pattern of having UI controls in a form that floats on top of the map/scene in the top right.

Now, Forms samples can, too. On phones, the form is shown pinned to the top of the window, taking the full width. On desktop and tablets (including iPad!), the behavior is very similar to WPF desktop.

This was used in the Display Drawing Status sample, which didn't have a good place for the status bar.