Automated Readme & Metadata json updates
Content was inspected for basic issues (formatting, spelling, language, basic accuracy) on Android; fixes were automatically applied to the four other platforms. Note that accuracy of content relative to the actual UI implementation was not verified. Before merging, the readme content should be compared to the actual implementation on all platforms for all samples.