Update sample metadata tools for VS
Previously, the solution templates were named with .sln extensions. This caused Visual Studio to detect them as solutions when using the git browser. To prevent this, the solutions have been renamed to .slntemplate and the script has been updated to rename them to .sln in the output. I've confirmed that the template solutions no longer appear in Visual Studio.